requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
findspark==2.0.1
fonttools==4.38.0
greenlet==2.0.2
idna==3.4
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prettytable==3.6.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.5
Pygments==2.14.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
spark==0.2.1
SQLAlchemy==2.0.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
